Update ZohoHTTPConnector.php
Throw an exception if an error ocurres, curl. 

Also you could add to the config, something like DEV_MODE, so in the line 55 you could add:

```php
if (DEV_MODE) {
     curl_setopt($curl_pointer, CURLOPT_SSL_VERIFYPEER, false);
}
```
It fails otherwise on local pc when you develop if your certificates are not correct saving a lot of time for the developers.